refactor: JpaRepository 구현, 메서드 수정
