+++ src/main/java/com/example/surveyapp/domain/product/infrastructure/ProductJpaRepository.java
@@ -1,4 +1,4 @@
-package com.example.surveyapp.domain.product.domain.repository;
+package com.example.surveyapp.domain.product.infrastructure;
 
 import com.example.surveyapp.domain.product.domain.model.Product;
 import com.example.surveyapp.domain.product.domain.model.Status;
@@ -8,8 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProductRepository {
-    
+public interface ProductJpaRepository extends JpaRepository<Product, Long> {
     Page<Product> findAllByStatusAndIsDeletedFalse(Status status, Pageable pageable);
 
     Optional<Product> findByIdAndStatusAndIsDeletedFalse(Long id, Status status);
